requirements.txt
asgiref==3.7.2
bcrypt==4.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
install==1.3.5
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.4.0
